pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "customtkinter>=5.2.2",
    "psutil>=7.1.2",
    "pyserial>=3.5",
]
